@@ -2,56 +2,86 @@
   "name": "nanogen",
   "version": "0.4.2",
   "description": "Minimalist static site generator in Node.js",
-  "repository": "doug2k1/nanogen",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/doug2k1/nanogen.git"
+  },
   "keywords": [
     "static site generator"
   ],
   "author": "Douglas Matoso <dmatoso@outlook.com> (https://dmatoso.com)",
   "license": "MIT",
-  "main": "./lib/index.js",
+  "type": "module",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": "./dist/index.js"
+  },
   "bin": {
-    "nanogen": "./lib/cli.js"
+    "nanogen": "dist/cli.js"
   },
   "files": [
-    "lib",
+    "dist",
     "template"
   ],
   "scripts": {
-    "test": "mocha --recursive",
-    "test:coverage": "nyc --reporter=html --reporter=text npm run test",
-    "test:ci": "nyc --reporter=lcov npm run test",
-    "prettier": "prettier \"./{lib,test}/**/*.js\"",
+    "build": "tsup-node",
+    "clean": "rimraf dist",
+    "typecheck": "tsc --noEmit",
+    "test": "vitest run",
+    "test:watch": "vitest watch",
+    "test:coverage": "vitest --coverage",
+    "prettier": "prettier \"./{src,docs-src,template}/**/*.{js,ts,md,css,html}\"",
     "prettier:check": "npm run prettier -- -l",
-    "lint": "eslint \"./{lib,test}/**/*.js\"",
-    "ci": "npm run lint && npm run prettier:check && npm run test:ci",
-    "site:start": "nanogen start",
-    "site:build": "cross-env NODE_ENV=production nanogen build"
+    "lint": "eslint \"./{src,docs-src,template}/**/*\"",
+    "start": "tsx src/cli.ts",
+    "start:docs": "node dist/cli start",
+    "build:docs": "cross-env NODE_ENV=production node dist/cli build",
+    "update-template": "tsx scripts/update-template-version.ts"
   },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=20.0.0"
   },
   "dependencies": {
-    "chalk": "^2.4.1",
-    "chokidar": "^2.0.4",
-    "ejs": "^2.6.1",
-    "front-matter": "^3.0.0",
-    "fs-extra": "^7.0.0",
-    "glob": "^7.1.3",
-    "live-server": "^1.2.0",
+    "chokidar": "^4.0.3",
+    "commander": "^13.0.0",
+    "ejs": "^3.1.10",
+    "front-matter": "^4.0.2",
+    "fs-extra": "^11.3.0",
+    "glob": "^11.0.1",
+    "live-server": "^1.2.2",
     "lodash.debounce": "^4.0.8",
-    "marked": "^0.5.1",
-    "meow": "^5.0.0",
-    "ora": "^3.0.0"
+    "marked": "^15.0.6",
+    "nanospinner": "^1.2.2",
+    "picocolors": "^1.1.1"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
-    "cross-env": "^5.2.0",
-    "eslint": "^5.6.1",
-    "eslint-config-prettier": "^3.1.0",
-    "eslint-plugin-node": "^7.0.1",
-    "mocha": "^5.2.0",
-    "nyc": "^13.0.1",
-    "prettier": "^1.14.3",
-    "sinon": "^6.3.5"
-  }
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.18.0",
+    "@tsconfig/node20": "^20.1.4",
+    "@types/ejs": "^3.1.5",
+    "@types/fs-extra": "^11.0.4",
+    "@types/live-server": "^1.2.3",
+    "@types/lodash.debounce": "^4.0.9",
+    "@types/meow": "^6.0.0",
+    "@types/node": "^22.10.6",
+    "@vitest/coverage-v8": "^2.1.8",
+    "cross-env": "^7.0.3",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-n": "^17.15.1",
+    "globals": "^15.14.0",
+    "nyc": "^17.1.0",
+    "prettier": "^3.4.2",
+    "rimraf": "^6.0.1",
+    "tsup": "^8.3.5",
+    "tsx": "^4.19.2",
+    "typescript": "^5.7.3",
+    "vite-tsconfig-paths": "^5.1.4",
+    "vitest": "^2.1.8"
+  },
+  "bugs": {
+    "url": "https://github.com/doug2k1/nanogen/issues"
+  },
+  "homepage": "https://github.com/doug2k1/nanogen#readme"
 }
